chore: rename component to match design system terminology
